+++ src/gamedata/fonts/v_font.h
@@ -114,7 +114,7 @@ protected:
 		const void *ranges, int total_colors, const PalEntry *palette);
 	void FixXMoves();
 
-	static int SimpleTranslation (uint8_t *colorsused, uint8_t *translation,
+	static int SimpleTranslation (uint32_t *colorsused, uint8_t *translation,
 		uint8_t *identity, TArray<double> &Luminosity);
 
 	void ReadSheetFont(TArray<FolderEntry> &folderdata, int width, int height, const DVector2 &Scale);
